requirements.txt
alembic==1.13.3
beautifulsoup4==4.12.3
fastapi==0.115.6
httpx==0.27.2
openai==1.55.3
psycopg[binary]==3.2.3
pydantic-settings==2.6.1
pypdf==5.1.0
PyYAML==6.0.2
qdrant-client==1.12.1
sentence-transformers==3.3.1
SQLAlchemy==2.0.36
uvicorn[standard]==0.32.1

# Test/dev
pytest==8.3.4
ruff==0.8.2
